+++ index.d.ts
@@ -33,6 +33,13 @@ export class Registry {
 	getSingleMetric(name: string): Metric;
 
 	/**
+	 * Set static labels to every metric emitted by this registry
+	 * @param object of name/value pairs:
+	 * { defaultLabel: "value", anotherLabel: "value 2" }
+	 */
+	setDefaultLabels(labels: Object): void;
+
+	/**
 	 * Get a string representation of a single metric by name
 	 * @param name The name of the metric
 	 */