+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 ### Added
+- Ability to set default labels by registry
 ### Changed
 - Convert code base to ES2015 code (node 4)
   - add engines field to package.json
@@ -13,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
   - Arrow functions over binding or putting `this` in a variable
   - Use template strings
   - `prototype` -> `class`
+
 
 ## [9.1.1] - 2017-06-17
 ### Changed